+++ motor.h
@@ -3,11 +3,11 @@
 
 class Motor {
     public:
-        Motor(PinName forward, PinName backward, int multiplier = 1);
+        Motor(PinName forward, PinName backward, float multiplier = 1);
         void set_speed(float speed);
         float get_speed();
     private:
         PwmOut forward;
         PwmOut backward;
-        int multiplier;
+        float multiplier;
 };
